+++ core/src/main/java/org/apache/metamodel/membrane/app/registry/TenantContext.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.membrane.app;
+package org.apache.metamodel.membrane.app.registry;
 
 /**
  * Represents a context-object containing all the information and services