+++ core/src/main/java/org/apache/metamodel/membrane/app/registry/cache/CachedDataSourceRegistryWrapper.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.membrane.app;
+package org.apache.metamodel.membrane.app.registry.cache;
 
 import java.util.List;
 import java.util.concurrent.TimeUnit;
@@ -26,6 +26,7 @@ import org.apache.metamodel.MetaModelException;
 import org.apache.metamodel.factory.DataContextProperties;
 import org.apache.metamodel.membrane.app.exceptions.DataSourceAlreadyExistException;
 import org.apache.metamodel.membrane.app.exceptions.NoSuchDataSourceException;
+import org.apache.metamodel.membrane.app.registry.DataSourceRegistry;
 import org.apache.metamodel.util.FileHelper;
 
 import com.google.common.cache.CacheBuilder;