+++ CHANGES.md
@@ -3,5 +3,6 @@
 * [METAMODEL-1153] - Added endpoints for deleting data sources.
 * [METAMODEL-1147] - Implemented Swagger codegen to build model classes from swagger file.
 * [METAMODEL-1149] - File based registry for tenant and data source information.
+* [METAMODEL-1166] - Caching of tenant and data source context information
 * [METAMODEL-1154] - Added update and delete capabilities in POST call to table data endpoint.
 * Established project source control, structure and build.